src/utils.py
import torch
from torchvision.transforms import Compose, ToTensor, Normalize
from torch.utils.data import Dataset, DataLoader
import os
from PIL import Image
import numpy as np
from math import log10

# Placeholder for data loading and preprocessing
class ImageDataset(Dataset):
    def __init__(self, root_dir, transform=None):
        self.root_dir = root_dir
        self.transform = transform
        self.input_images = sorted([os.path.join(root_dir, 'Input', img) for img in os.listdir(os.path.join(root_dir, 'Input'))])
        self.gt_images = sorted([os.path.join(root_dir, 'GT', img) for img in os.listdir(os.path.join(root_dir, 'GT'))])

    def __len__(self):
        return len(self.input_images)

    def __getitem__(self, idx):
        input_img_path = self.input_images[idx]
        gt_img_path = self.gt_images[idx]

        input_image = Image.open(input_img_path).convert('RGB')
        gt_image = Image.open(gt_img_path).convert('RGB')

        if self.transform:
            input_image = self.transform(input_image)
            gt_image = self.transform(gt_image)

        return input_image, gt_image

def get_dataloader(data_dir, batch_size, train=True):
    transform = Compose([
        ToTensor(),
        # Normalize(mean=[0.5, 0.5, 0.5], std=[0.5, 0.5, 0.5]) # Example normalization
    ])
    dataset = ImageDataset(data_dir, transform=transform)
    return DataLoader(dataset, batch_size=batch_size, shuffle=train)

# Placeholder for evaluation metrics
def calculate_psnr(img1, img2):
    mse = torch.mean((img1 - img2) ** 2)
    if mse == 0:
        return 100
    PIXEL_MAX = 1.0 # Assuming images are normalized to [0, 1]
    return 20 * log10(PIXEL_MAX / torch.sqrt(mse))

def calculate_ssim(img1, img2):
    # Placeholder for SSIM calculation
    # This would typically involve a more complex implementation or a library like scikit-image
    return 0.0 # Dummy value

def calculate_mse(img1, img2):
    return torch.mean((img1 - img2) ** 2)